refactor models into file components
